pendantd/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "pendantd"
version = "0.1.0.dev0"
description = "Pi-side service for robot-md-pendant"
requires-python = ">=3.11"
license = "Apache-2.0"
dependencies = [
  "websockets>=12.0",
  "jsonschema>=4.21",
  "pyyaml>=6.0",
  "watchfiles>=0.21",
  "claude-agent-sdk>=0.1.0",
  "openwakeword>=0.6",
]

[project.optional-dependencies]
dev = ["pytest>=8", "pytest-asyncio>=0.23"]

[project.scripts]
pendantd = "pendantd.__main__:main"

[tool.hatch.build.targets.wheel]
packages = ["src/pendantd"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
